requirements2.txt
appdirs==1.4.4
beautifulsoup4==4.12.0
boto3==1.26.97
botocore==1.29.97
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
frozendict==2.3.6
html5lib==1.1
idna==3.4
jmespath==1.0.1
joblib==1.2.0
lxml==4.9.2
multitasking==0.0.11
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
pandas-datareader==0.10.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7.1
regex==2023.3.22
requests==2.28.2
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.65.0
urllib3==1.26.15
vaderSentiment==3.3.2
webencodings==0.5.1
yfinance==0.2.13
